refactor(cli): extract workflow gate handler
